Use Locale instead of String, move package
