+++ src/main/java/net/minestom/server/network/packet/client/common/ClientSettingsPacket.java
@@ -1,9 +1,9 @@
 package net.minestom.server.network.packet.client.common;
 
-import net.minestom.server.entity.ClientSettings;
 import net.minestom.server.network.NetworkBuffer;
 import net.minestom.server.network.NetworkBufferTemplate;
 import net.minestom.server.network.packet.client.ClientPacket;
+import net.minestom.server.network.player.ClientSettings;
 import org.jetbrains.annotations.NotNull;
 
 public record ClientSettingsPacket(@NotNull ClientSettings settings) implements ClientPacket {
